+++ src/main/java/net/telepathicgrunt/bumblezone/items/EmptyBucketDispenseBehavior.java
@@ -14,6 +14,7 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Position;
 import net.minecraft.world.World;
 import net.telepathicgrunt.bumblezone.blocks.BzBlocks;
+import net.telepathicgrunt.bumblezone.mixin.ItemDispenserBehaviorInvoker;
 
 
 public class EmptyBucketDispenseBehavior extends ItemDispenserBehavior {
@@ -43,7 +44,15 @@ public class EmptyBucketDispenseBehavior extends ItemDispenserBehavior {
             else
                 stack = new ItemStack(BzItems.SUGAR_WATER_BUCKET);
         } else {
-            return DEFAULT_EMPTY_BUCKET_DISPENSE_BEHAVIOR.dispense(source, stack);
+            // If it instanceof DefaultDispenseItemBehavior, call dispenseStack directly to avoid
+            // playing particles and sound twice due to dispense method having that by default.
+            if(DEFAULT_EMPTY_BUCKET_DISPENSE_BEHAVIOR instanceof ItemDispenserBehavior) {
+                return ((ItemDispenserBehaviorInvoker)DEFAULT_EMPTY_BUCKET_DISPENSE_BEHAVIOR).invokeDispenseSilently(source, stack);
+            }
+            else {
+                // Fallback to dispense as someone chose to make a custom class without dispenseStack.
+                return DEFAULT_EMPTY_BUCKET_DISPENSE_BEHAVIOR.dispense(source, stack);
+            }
         }
 
         return stack;