+++ src/main/java/net/telepathicgrunt/bumblezone/items/GlassBottleDispenseBehavior.java
@@ -20,6 +20,7 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Position;
 import net.telepathicgrunt.bumblezone.blocks.BzBlocks;
 import net.telepathicgrunt.bumblezone.blocks.HoneycombBrood;
+import net.telepathicgrunt.bumblezone.mixin.ItemDispenserBehaviorInvoker;
 
 
 public class GlassBottleDispenseBehavior extends ItemDispenserBehavior {
@@ -84,7 +85,15 @@ public class GlassBottleDispenseBehavior extends ItemDispenserBehavior {
                 stack = new ItemStack(BzItems.SUGAR_WATER_BOTTLE);
         }
         else {
-            return DEFAULT_GLASS_BOTTLE_DISPENSE_BEHAVIOR.dispense(source, stack);
+            // If it instanceof DefaultDispenseItemBehavior, call dispenseStack directly to avoid
+            // playing particles and sound twice due to dispense method having that by default.
+            if(DEFAULT_GLASS_BOTTLE_DISPENSE_BEHAVIOR instanceof ItemDispenserBehavior) {
+                return ((ItemDispenserBehaviorInvoker)DEFAULT_GLASS_BOTTLE_DISPENSE_BEHAVIOR).invokeDispenseSilently(source, stack);
+            }
+            else {
+                // Fallback to dispense as someone chose to make a custom class without dispenseStack.
+                return DEFAULT_GLASS_BOTTLE_DISPENSE_BEHAVIOR.dispense(source, stack);
+            }
         }
 
         return stack;